The Game class is Created
